# backend/openapi/model_login_response.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Proof Pass API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.1
 */

package openapi




type LoginResponse struct {

	Token string `json:"token,omitempty"`
}

// AssertLoginResponseRequired checks if the required fields are not zero-ed
func AssertLoginResponseRequired(obj LoginResponse) error {
	return nil
}

// AssertLoginResponseConstraints checks if the values respects the defined constraints
func AssertLoginResponseConstraints(obj LoginResponse) error {
	return nil
}