# backend/openapi/model_user_login.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Proof Pass API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.1
 */

package openapi




type UserLogin struct {

	Email string `json:"email,omitempty"`

	Code string `json:"code,omitempty"`
}

// AssertUserLoginRequired checks if the required fields are not zero-ed
func AssertUserLoginRequired(obj UserLogin) error {
	return nil
}

// AssertUserLoginConstraints checks if the values respects the defined constraints
func AssertUserLoginConstraints(obj UserLogin) error {
	return nil
}