# openapi/v1/ticketing.yaml
openapi: 3.0.3
info:
  title: Proof Pass API
  version: 0.0.1
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /user/request-verification-code:
    post:
      summary: Request an email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserEmailVerificationRequest"
      responses:
        "200":
          description: Verification code sent successfully
  /user/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserEmailVerificationRequest:
      type: object
      properties:
        email:
          type: string
    UserLogin:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
...